# docs/openapi.yaml
openapi: 3.0.0
info:
  contact:
    email: msmariaulfa33@gmail.com
  version: 0.0.1
  title: DNR Service Swagger Docs
  description: Service for DNR Website
servers:
  - description: Local development
    url: 'http://localhost:3001'
...
paths:
...
  /api/v1/payment_terms:
    get:
      description: Get All Payment Terms
      tags:
        - payment-terms
      summary: List all Payment Terms
      operationId: getPaymentTerms
      security:
        - ApiKeyAuth: []
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Payment-terms'
        400:
          description: example error 400 response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error400Response'
        500:
          description: example error 500 response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error500Response'
        404:
          description: example error 404 response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error404Response'
    post:
      description: Find payment term by filtering
      tags:
        - payment-terms
      summary: Find payment term by filtering
      operationId: filterPaymentTerm
      security:
        - ApiKeyAuth: []
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Payment-terms'
        400:
          description: example error 400 response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error400Response'
        500:
          description: example error 500 response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error500Response'
        404:
          description: example error 404 response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error404Response'
        201:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Payment-terms'
  /api/v1/payment_terms/create:
    post:
      description: Create a payment term
      tags:
        - payment-terms
      summary: Create a term of payment
      operationId: createPaymentTerm
      security:
        - ApiKeyAuth: []
      responses:
        201:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Payment-terms'
  /api/v1/payment_terms/{id}:
    get:
      description: Get Payment Terms By Id
      summary: Get Payment Terms By Id
      operationId: getPaymentTermsById
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Payment-terms'
        400:
          description: example error 400 response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error400Response'
        500:
          description: example error 500 response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error500Response'
        404:
          description: example error 404 response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error404Response'
      tags:
        - payment-terms
    put:
      summary: Update payment term
      operationId: updatePaymentTerm
      security:
        - ApiKeyAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Payment-terms-update'
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Payment-terms'
        404:
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error404Response'
        400:
          description: example error 400 response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error400Response'
        500:
          description: example error 500 response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error500Response'
    patch:
      summary: Change payment term status
      operationId: changePaymentTerm
      security:
        - ApiKeyAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Payment-terms-change'
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Payment-terms'
        404:
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error404Response'
        400:
          description: example error 400 response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error400Response'
        500:
          description: example error 500 response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error500Response'
  /api/v1/payment_terms/active:
    get:
      description: Get All Payment Terms
      tags:
        - payment-terms
      summary: List all Payment Terms
      operationId: getPaymentTerms
      security:
        - ApiKeyAuth: []
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Payment-terms'
        400:
          description: example error 400 response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error400Response'
        500:
          description: example error 500 response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error500Response'
        404:
          description: example error 404 response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error404Response'
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-auth-token
  schemas:
...
    Error400Response:
      type: object
      required:
        - errorCode
        - message
      properties:
        errorCode:
          type: string
          enum:
            - INVALID_API_KEY
            - API_VALIDATION_ERROR
            - CAN_NOT_REGISTER
            - CAN_NOT_LOGIN
            - UNAUTHORIZED_USER
            - OUTLET_CREATION_ERROR
            - CREATE_ORDER_ERROR
            - ADDRESS_NOT_FOUND
            - ADD_TO_CART_ERROR
            - UPDATE_ORDER_ERROR
            - COMPLETE_PAYMENT_ORDER_ERROR
            - CANCEL_ORDER_ERROR
            - COMPLETE_ORDER_ERROR
            - FLASH_SALE_ERROR
            - EXISTING_CUSTOMER_ID
            - RESET_PASSWORD_ERROR
            - NEWS_ERROR
            - OUTLET_TYPES_ERROR
            - PROMOTION_PRODUCTS_EXIST
            - PROMOTION_CODE_EXIST
        message:
          type: string
        errors:
          $ref: '#/components/schemas/ArrayOfErrors'
    Error404Response:
      type: object
      properties:
        errorCode:
          type: string
          enum:
            - NOT_FOUND_ERROR
            - USER_NOT_FOUND_ERROR
            - PRODUCT_NOT_FOUND_ERROR
            - ORDER_NOT_FOUND_ERROR
            - BANK_ACCOUNT_NOT_FOUND_ERROR
            - CART_NOT_FOUND_ERROR
            - VERIFICATION_EMAIL_ERROR
            - NOTIFICATION_NOT_FOUND_ERROR
            - FLASH_SALE_NOT_FOUND_ERROR
            - NEWS_ERROR
            - OUTLET_TYPES_ERROR
            - PROMOTION_NOT_FOUND_ERROR
        message:
          type: string
        errors:
          $ref: '#/components/schemas/ArrayOfErrors'
      required:
        - errorCode
        - message
    Error500Response:
      type: object
      required:
        - errorCode
        - message
      properties:
        errorCode:
          type: string
          enum:
            - INTERNAL_SERVER_ERROR
        message:
          type: string
        errors:
          $ref: '#/components/schemas/ArrayOfErrors'
...
    ArrayOfErrors:
      nullable: true
      type: array
      items:
        oneOf:
          - type: string
          - type: object
...
    Payment-terms:
      type: object
      properties:
        type:
          type: string
        name:
          type: string
        days_due:
          type: number
        status:
          type: string
        created:
          type: string
          format: date-time
        updated:
          type: string
          format: date-time
...
    Payment-terms-update:
      type: object
      properties:
        name:
          type: string
        days_due:
          type: number
    Payment-terms-change:
      type: object
      properties:
        status:
          type: string
...